Add testenv command for CI pre/post processing
When the integration tests ran and failed there could be resources left
on the target system causing next run to fail.

This change adds a new internal command, testenv, that runs before and
after the CI test. Before it ensures that there are no resources around
that will fail the tests, and after it tries to remove resources left by
failed tests.