Move setting of default log level out of NewClient (#58)
If a new client was created via NewClient and the log level wasn't set
using env var, it wasn't possible to provide a different level than the
default, warn. The NewClient call itself also logs at Info level so the
only way to see those logs were to use env override.

We now only override the provided loggers level if the env var is set.
This change also provides a NewStandardLogger() function that returns a
StandardLogger with the default level set (Warn).