Remove Azure AD Graph dependency (#204) (#211)
If the app name wasn't provided when initializing a service principal
the SDK tried to look it up using the Azure AD Graph API. This API is
now being shut down by Microsoft. This primarily happened when the
service principal was initialized with JSON data created by the az
tool and the --sdk-auth command line option.

Instead of looking up the app name in Azure, we simply generate a name
from the app and tenant IDs.

Co-authored-by: Johan Andersson <johan.andersson@rubrik.com>